Update ipresolve for puppet7
(Backport changes from wikimedia)
